More Player and Boat Control Fixes
Smoothstep implemented for camera changes.
Switching requires the player to interact with the door.
(Could leave raycast in for boat view, but waste of processing.)
Cabin and chimneys added to the boat.
Steam particle effects added to cabin chimneys.
Look rotation etc. made less jittery.
Player no longer rotates with boat rotation (wrongly as in last commit)
when free roaming.
Player constantly faces the door when controlling the boat. Make the
player's renderers disable in future to give the impression of getting
into and out of the cabin.