Temple Section, Crate & Fixes
Added temple section and inhabited temple section.
Added crate model and made it spin, float, and destroy on contact with
the boat.
Fixed bug where player's local y value changes when exiting the door,
causing the player to get higher or lower and potentially allowing the
player to leave the boat.